requirements.txt
librosa==0.10.2.post1
matplotlib==3.8.4
moviepy==1.0.3
numpy==2.3.2
opencv_python==4.9.0.80
Pillow==11.3.0
PyYAML==6.0.1
PyYAML==6.0.2
scipy==1.16.1
torch==1.12.0
torch==2.2.2
tqdm==4.65.2
